Decrease startup delay by using lazy initialization
The LocalSettings.IsResetScrollPosition property is read during
construction of the main form. This forces the LocalSettings class
to be initialized. Because the LocalSettings class contains
non-trivial members, this takes a considerable amount of time.

However, the non-trivial members in LocalSettings are not used
during initial startup, so by changing them to be initialized lazily
a considerable startup delay (roughly one second on my machine) can
be avoided.